+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [19.18.0] - 2024-04-24
+
+## Added
+
+- Animation to email field when the email is valid
+
+## [19.17.0] - 2024-04-24
+
+## Added
+
+- background icon prop to Image Field
+
 ## [19.16.2] - 2024-04-18
 
 ## Added